+++ src/ui_basic/textarea.cc
@@ -125,7 +125,7 @@ void Textarea::set_fixed_width(int w) {
 void Textarea::draw(RenderTarget& dst) {
 	if (!text_.empty()) {
 		// Blit on pixel boundary (not float), so that the text is blitted pixel perfect.
-		Vector2f anchor(
+		Vector2i anchor(
 		   (align_ == Align::kCenter) ? get_w() / 2 : (align_ == UI::Align::kRight) ? get_w() : 0, 0);
 		dst.blit(anchor, rendered_text_->texts[0]->image, BlendMode::UseAlpha, align_);
 	}
